requirements.txt
telethon>=1.24.0,<2.0.0
aiohttp>=3.8.0,<4.0.0
python-dotenv>=0.19.0,<2.0.0
aiogram>=3.0.0,<4.0.0
